Add meal sharing feature and necessary libraries
The commit introduces a new feature to share meals which involves the creation of the lib/actions.js and the update of lib/meals.js to include a new function 'saveMeal'. This function saves meals to the database and writes images to the file system. Necessary libraries are also added to the package.json file like 'slugify', 'xss' and applying the appropriate changes in package-lock.json file. The UI updates corresponding to the new functionalities are also reflected in app/meals/share/page.js.